Add user hint about prescalers to the config management panels.
Loading a configuration has the side effect of processing all records.
Setting the frequency of the fractional synchronizer (even if it is set
to the same frequency as before) causes the phase of the prescaler to
jump, so when loading a configuration, the phase of the prescalers
(w.r.t. other EVRs) changes and the “synchronize prescalers” event has
to be sent in order to fix the phase again.