Avoid aborting initialization of output records.
When a device is temporarily unvailable when the IOC is started, the
values of output records cannot be read from the device, and thus these
values cannot be initialized.

However, this should not keep the record initialization routine from
completing, so that the record can be processed at a later time.

Therefore, we now catch such exceptions and print an error message but
continue with initializing the record instead of destroying the device
support instance for the record.