Use Python for generating .db and .req files.
The old Bash script have reached the point where adding new devices or
features became very hard. In particular, adding support for the EVM was
barely possible.

With this commit, the logic is completely reimplemented using Python,
but the generated files are mostly kept as they were before. The main
differences are:

- Some records are ordered a bit differently.
- Bugs (wrong addresses) in some of the records for the SFP module have
  been fixed.
- Bugs that caused “WriteAll” not to actually link to all output records
  have been fixed.