Generate records for CML logic for each output.
Instead of having one long file that explicitly lists the records for
all outputs, we now have a template file and generate the relevant
records for each output.

This commit also fixes the problem that the records for outputs with GTX
logic where not processed as part of the “write all” action.