Allow output redirection and improve error logging.
This commit enables output redirection in the IOC shell for all the IOC
shell functions that are provided by this device support.

It also improves the error logging function to use the errlog facility
provided by EPICS Base whenever possible. And to also respect output
redirections which are in force for the current thread.

Closes #12.