Add early, preliminary support for the VME-EVM-300.
This implementation is far from complete and is still missing records
for a number of crucial settings.

As part of this change, the documentation is restructured to use
separate documents for the process variables of the now three device
types. This change has been made because the single document that was
used previously was getting too long.